+++ main.go
@@ -88,6 +88,7 @@ var (
 		CacheTimeCatchAll:            10 * time.Minute,
 		SQLIndexerDatabaseFolderPath: "/tmp/", // TODO: Another default directory?
 		SearchCacheSize:              100,
+		SearchCacheTTL:               10 * time.Minute,
 	}
 )
 
@@ -125,8 +126,9 @@ type Config struct {
 	CacheTimeSearch              time.Duration `config:"cache_time.search"`
 	CacheTimeCategories          time.Duration `config:"cache_time.categories"`
 	CacheTimeCatchAll            time.Duration `config:"cache_time.catch_all"`
-	SQLIndexerDatabaseFolderPath string        `config:"sql_indexer.database_folder_path"`
-	SearchCacheSize              int           `config:"search.cache_size"`
+	SQLIndexerDatabaseFolderPath string        `config:"sql_indexer.database_folder_path"` // technical preview, used by the SQL storage indexer
+	SearchCacheSize              int           `config:"search.cache_size"`                // technical preview, used by the SQL storage indexer
+	SearchCacheTTL               time.Duration `config:"search.cache_ttl"`                 // technical preview, used by the SQL storage indexe^
 }
 
 func main() {
@@ -200,9 +202,9 @@ func main() {
 		defer fakeServer.Stop()
 	}
 
-	var searchCache *expirable.LRU[string, string]
+	var searchCache *expirable.LRU[string, []byte]
 	if featureSQLStorageIndexer && featureEnableSearchCache {
-		searchCache = expirable.NewLRU[string, string](config.SearchCacheSize, nil, config.CacheTimeSearch)
+		searchCache = expirable.NewLRU[string, []byte](config.SearchCacheSize, nil, config.SearchCacheTTL)
 	}
 
 	indexer := initIndexer(ctx, logger, apmTracer, config, searchCache)
@@ -245,7 +247,6 @@ func initDatabase(ctx context.Context, logger *zap.Logger, databaseFolderPath, d
 	if exists {
 		err = os.Remove(dbPath)
 		if err != nil {
-			logger.Fatal("failed to delete previous database", zap.String("path", dbPath), zap.Error(err))
 			return nil, fmt.Errorf("failed to delete previous database (path %q): %w", dbPath, err)
 		}
 	}
@@ -333,7 +334,7 @@ func initMetricsServer(logger *zap.Logger) {
 	}()
 }
 
-func initIndexer(ctx context.Context, logger *zap.Logger, apmTracer *apm.Tracer, config *Config, cache *expirable.LRU[string, string]) Indexer {
+func initIndexer(ctx context.Context, logger *zap.Logger, apmTracer *apm.Tracer, config *Config, cache *expirable.LRU[string, []byte]) Indexer {
 	tx := apmTracer.StartTransaction("initIndexer", "backend.init")
 	defer tx.End()
 
@@ -379,7 +380,7 @@ func initStorageIndexer(ctx context.Context, logger *zap.Logger, apmTracer *apm.
 	}), nil
 }
 
-func initSQLStorageIndexer(ctx context.Context, logger *zap.Logger, apmTracer *apm.Tracer, config *Config, cache *expirable.LRU[string, string]) (*internalStorage.SQLIndexer, error) {
+func initSQLStorageIndexer(ctx context.Context, logger *zap.Logger, apmTracer *apm.Tracer, config *Config, cache *expirable.LRU[string, []byte]) (*internalStorage.SQLIndexer, error) {
 	storageClient, err := newStorageClient(ctx, logger)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create storage client: %w", err)
@@ -417,7 +418,7 @@ func newStorageClient(ctx context.Context, logger *zap.Logger) (*gstorage.Client
 	return gstorage.NewClient(ctx, opts...)
 }
 
-func initServer(logger *zap.Logger, apmTracer *apm.Tracer, config *Config, indexer Indexer, cache *expirable.LRU[string, string]) *http.Server {
+func initServer(logger *zap.Logger, apmTracer *apm.Tracer, config *Config, indexer Indexer, cache *expirable.LRU[string, []byte]) *http.Server {
 	router := mustLoadRouter(logger, config, indexer, cache)
 	apmgorilla.Instrument(router, apmgorilla.WithTracer(apmTracer))
 
@@ -475,6 +476,25 @@ func getConfig(logger *zap.Logger) (*Config, error) {
 	if err != nil {
 		return nil, fmt.Errorf("unpacking config failed (path: %s): %w", configPath, err)
 	}
+
+	// Parse environment variables to override config values in technical preview mode
+	if path, found := os.LookupEnv("EPR_SQL_INDEXER_DATABASE_FOLDER_PATH"); found && path != "" {
+		config.SQLIndexerDatabaseFolderPath = path
+	}
+	if v, found := os.LookupEnv("EPR_SQL_INDEXER_SEARCH_CACHE_SIZE"); found && v != "" {
+		cacheSize, err := strconv.Atoi(v)
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse EPR_SQL_INDEXER_SEARCH_CACHE_SIZE environment variable: %w", err)
+		}
+		config.SearchCacheSize = cacheSize
+	}
+	if v, found := os.LookupEnv("EPR_SQL_INDEXER_SEARCH_CACHE_TTL"); found && v != "" {
+		cacheTTL, err := time.ParseDuration(v)
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse EPR_SQL_INDEXER_SEARCH_CACHE_TTL environment variable: %w", err)
+		}
+		config.SearchCacheTTL = cacheTTL
+	}
 	return &config, nil
 }
 
@@ -491,8 +511,9 @@ func printConfig(logger *zap.Logger, config *Config) {
 	logger.Info("Cache time for /search: " + config.CacheTimeSearch.String())
 	logger.Info("Cache time for /categories: " + config.CacheTimeCategories.String())
 	logger.Info("Cache time for all others: " + config.CacheTimeCatchAll.String())
-	logger.Info("Database path: " + config.SQLIndexerDatabaseFolderPath)
-	logger.Info("LRU cache size (search requests): " + strconv.Itoa(config.SearchCacheSize))
+	logger.Info("(technical preview) SQL storage indexer database path: " + config.SQLIndexerDatabaseFolderPath)
+	logger.Info("(technical preview) Search cache size (SQL storage indexer): " + strconv.Itoa(config.SearchCacheSize))
+	logger.Info("(technical preview) Search cache TTL (SQL storage indexer): " + config.SearchCacheTTL.String())
 }
 
 func ensurePackagesAvailable(ctx context.Context, logger *zap.Logger, indexer Indexer) {
@@ -516,7 +537,7 @@ func ensurePackagesAvailable(ctx context.Context, logger *zap.Logger, indexer In
 	metrics.NumberIndexedPackages.Set(float64(len(packages)))
 }
 
-func mustLoadRouter(logger *zap.Logger, config *Config, indexer Indexer, cache *expirable.LRU[string, string]) *mux.Router {
+func mustLoadRouter(logger *zap.Logger, config *Config, indexer Indexer, cache *expirable.LRU[string, []byte]) *mux.Router {
 	router, err := getRouter(logger, config, indexer, cache)
 	if err != nil {
 		logger.Fatal("failed go configure router", zap.Error(err))
@@ -524,7 +545,7 @@ func mustLoadRouter(logger *zap.Logger, config *Config, indexer Indexer, cache *
 	return router
 }
 
-func getRouter(logger *zap.Logger, config *Config, indexer Indexer, cache *expirable.LRU[string, string]) (*mux.Router, error) {
+func getRouter(logger *zap.Logger, config *Config, indexer Indexer, cache *expirable.LRU[string, []byte]) (*mux.Router, error) {
 	if featureProxyMode {
 		logger.Info("Technical preview: Proxy mode is an experimental feature and it may be unstable.")
 	}